Create a MenuHost with new behaviors
